refactor: 계층 구조를 생각하여 domain, entity, dto 클래스를 최대한 서로 의존하지 않을 수 있도록 수정
